+++ weblate/formats/models.py
@@ -48,8 +48,7 @@ class FileFormatLoader(ClassLoader):
                 fileformat.get_class()
             except (AttributeError, ImportError) as error:
                 result.pop(fileformat.format_id)
-                if fileformat.format_id != "rc":
-                    self.errors[fileformat.format_id] = str(error)
+                self.errors[fileformat.format_id] = str(error)
 
         return result
 
@@ -88,13 +87,13 @@ class FormatsConf(AppConf):
         "weblate.formats.ttkit.AdvSubStationAlphaFormat",
         "weblate.formats.ttkit.SubStationAlphaFormat",
         "weblate.formats.ttkit.DTDFormat",
-        "weblate.formats.ttkit.WindowsRCFormat",
         "weblate.formats.ttkit.FlatXMLFormat",
         "weblate.formats.external.XlsxFormat",
         "weblate.formats.txt.AppStoreFormat",
         "weblate.formats.convert.HTMLFormat",
         "weblate.formats.convert.IDMLFormat",
         "weblate.formats.convert.OpenDocumentFormat",
+        "weblate.formats.convert.WindowsRCFormat",
     )
 
     class Meta: